Still Rocking It
Made it through Evil Pleasures but just remembered the error in condition 5 where it pretends there isn't a user the action is being applied to when there definitely is. Evil Pleasures is correct, but the ones before it are most likely not. Other than that, ready to keep moving forward.